src/edu/cscc/RPSLSpock.java
package edu.cscc;
import java.util.Random;

/**
 * Playing rock, paper, scissors, lizard, spock
 * @author Mutasem Alhariri 10/13/2019
 */

public class RPSLSpock {

    static Random rand = new Random(System.currentTimeMillis());

    public static final String ROCK = "rock";
    public static final String PAPER = "paper";
    public static final String SCISSORS = "scissors";
    public static final String LIZARD = "lizard";
    public static final String SPOCK = "spock";

    /**
     * Validates the value of the user's pick
     * @param pick String user's pick
     * @return true if pick value is equal to the value of one of these strings: ROCK, PAPER, SCISSORS, LIZARD,
     *         or SPOCK. otherwise return false, and user's pick cannot be null.
     */

    public static boolean isValidPick(String pick) {
        if (pick == null) {
            return false;
        }
        pick = pick.trim();
        return (ROCK.equalsIgnoreCase(pick) ||
                PAPER.equalsIgnoreCase(pick) ||
                SCISSORS.equalsIgnoreCase(pick) ||
                LIZARD.equalsIgnoreCase(pick) ||
                SPOCK.equalsIgnoreCase(pick));
    }

    /**
     * Generates computer's pick by generating a random number between 1 and 5, then assigns the value of ROCK,
     *        PAPER, SCISSORS, LIZARD, OR SPOCK TO the string pick.
     * @return
     */

    public static String generatePick() {
        String pick = null;
        switch (rand.nextInt(5)) {
            case 0:
                pick = ROCK;
                break;
            case 1:
                pick = PAPER;
                break;
            case 2:
                pick = SCISSORS;
                break;
            case 3:
                pick = LIZARD;
                break;
            case 4:
                pick = SPOCK;
                break;
        }
        return pick;
    }

    /**
     * Checks if computer wins by comparing c_pick with h_pick
     * @param c_pick String, computer pick
     * @param h_pick String, human's pick
     * @return true if the computer wins otherwise return false
     */

    public static boolean isComputerWin(String c_pick,String h_pick) {
        h_pick = h_pick.toLowerCase();
        return ((ROCK.equals(c_pick) && (SCISSORS.equals(h_pick) || LIZARD.equals(h_pick))) ||
                (PAPER.equals(c_pick) && (ROCK.equals(h_pick) || SPOCK.equals(h_pick))) ||
                (SCISSORS.equals(c_pick) && (PAPER.equals(h_pick) || LIZARD.equals(h_pick))) ||
                (LIZARD.equals(c_pick) && (PAPER.equals(h_pick) || SPOCK.equals(h_pick))) ||
                (SPOCK.equals(c_pick) && (ROCK.equals(h_pick) || SCISSORS.equals(h_pick))));
    }
}